refactor : 프로젝트 폴더구조 리팩터링
