+++ src/main/java/com/web/flower/security/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.web.flower.security;
+package com.web.flower.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
