--- src/main/java/com/web/flower/security/repository/RefreshTokenRepository.java
@@ -1,14 +1,0 @@
-package com.web.flower.security.repository;
-
-import com.web.flower.security.domain.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, UUID> {
-
-    Optional<RefreshToken> findById(UUID id);
-
-    void deleteById(UUID id);
-}
